# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trademark Ledger & Case Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases/{folderId}/stages:
    get:
      operationId: listCaseStages
      tags: [stages]
      summary: Get all stages for a case
      parameters:
        - name: folderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: All stages for the case
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CaseStage" }
  /cases/{folderId}/stages/{stageNumber}:
    get:
      operationId: getCaseStage
      tags: [stages]
      summary: Get a specific stage
      parameters:
        - name: folderId
          in: path
          required: true
          schema: { type: string }
        - name: stageNumber
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Stage detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CaseStage" }
    patch:
      operationId: updateCaseStage
      tags: [stages]
      summary: Update a stage (status, sub-status, timeline, etc.)
      parameters:
        - name: folderId
          in: path
          required: true
          schema: { type: string }
        - name: stageNumber
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StageUpdate" }
      responses:
        "200":
          description: Updated stage
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CaseStage" }

  # ── STAGE 2 ASSIGNMENTS ──────────────────────────────────────────────────
...
components:
  schemas:
...
    StageTimelineEvent:
      type: object
      required: [date, description]
      properties:
        date: { type: string }
        description: { type: string }
    StageHistoryEntry:
      type: object
      required: [changedAt, field, oldValue, newValue]
      properties:
        changedAt: { type: string }
        field: { type: string }
        oldValue: { type: ["string", "null"] }
        newValue: { type: ["string", "null"] }
    CaseStage:
      type: object
      required: [id, caseId, stageNumber, createdAt, updatedAt]
      properties:
        id: { type: integer }
        caseId: { type: integer }
        stageNumber: { type: integer }
        status: { type: ["string", "null"] }
        subStatus: { type: ["string", "null"] }
        timeline:
          type: array
          items: { $ref: "#/components/schemas/StageTimelineEvent" }
        history:
          type: array
          items: { $ref: "#/components/schemas/StageHistoryEntry" }
        createdAt: { type: string }
        updatedAt: { type: string }
    StageUpdate:
      type: object
      properties:
        status: { type: string }
        subStatus: { type: string }
        timelineEvent:
          $ref: "#/components/schemas/StageTimelineEvent"

    # ── ASSIGNMENT ───────────────────────────────────────────────────────────
...